+++ selvad/include/selva_io.h
@@ -24,6 +24,41 @@ struct SelvaDbVersionInfo {
     __nonstring char updated_with[SELVA_DB_VERSION_SIZE];
 };
 
+/**
+ * Replication mode.
+ */
+enum replication_mode {
+    SELVA_REPLICATION_MODE_NONE = 0,
+    SELVA_REPLICATION_MODE_ORIGIN,
+    SELVA_REPLICATION_MODE_REPLICA,
+};
+
+/**
+ * Marks an EID used for SDB.
+ * Used to distinguish between an SDB and command, which is useful
+ * with data structures that can contain both using the same pointers.
+ */
+#define EID_MSB_MASK (~(~(typeof(uint64_t))0 >> 1))
+
+/**
+ * Absolute save/load order for the dump files.
+ * The data is not annotated in any way in the file/stream, i.e. there is no
+ * automatic metadata of any sort. Compare reading an SDB dump to loading from
+ * a tape. Every piece of code must know exactly what's being loaded next and
+ * when its own data ends.
+ */
+enum selva_io_load_order {
+    SELVA_IO_ORD_HIERARCHY = 0,
+    NR_SELVA_IO_ORD,
+};
+
+struct selva_io_serializer {
+    bool (*is_ready)(); /*!< Is ready to be serialized. */
+    int (*deserialize)(struct selva_io *io);
+    void (*serialize)(struct selva_io *io);
+    void (*flush)(void);
+};
+
 enum selva_io_flags {
     SELVA_IO_FLAGS_READ = 0x0001, /*!< This is a read op. */
     SELVA_IO_FLAGS_WRITE = 0x0002, /*!< This is a write op. */
@@ -86,12 +121,28 @@ struct selva_io {
 };
 #endif
 
+enum selva_io_dump_state {
+    SELVA_DB_DUMP_NONE = 0x00, /*!< No dump operation running. */
+    SELVA_DB_DUMP_ACTIVE_CHILD = 0x01, /*!< There is an active child. */
+    SELVA_DB_DUMP_IS_CHILD = 0x02, /*!< This is the child process. */
+};
+
 #define SELVA_IO_FLAGS_MODE_MASK (SELVA_IO_FLAGS_READ | SELVA_IO_FLAGS_WRITE)
 
 /**
  * Get the version info.
  */
 SELVA_IO_EXPORT(void, selva_io_get_ver, struct SelvaDbVersionInfo *nfo);
+
+/**
+ * Register an SDB serializer.
+ * @param serializer A pointer to the struct is not held.
+ */
+SELVA_IO_EXPORT(void, selva_io_register_serializer, enum selva_io_load_order ord, const struct selva_io_serializer *serializer);
+
+SELVA_IO_EXPORT(void, selva_io_set_dirty, void);
+
+SELVA_IO_EXPORT (enum selva_io_dump_state, selva_io_get_dump_state, void);
 
 /**
  * Open the last good SDB for reading.
@@ -147,8 +198,29 @@ SELVA_IO_EXPORT(double, selva_io_load_double, struct selva_io *io);
 SELVA_IO_EXPORT(const char*, selva_io_load_str, struct selva_io *io, size_t *len);
 SELVA_IO_EXPORT(struct selva_string *, selva_io_load_string, struct selva_io *io);
 
+/**
+ * Get the replication mode.
+ */
+SELVA_IO_EXPORT(enum replication_mode, selva_replication_get_mode, void);
+
+/**
+ * Replicate a command buffer to replicas.
+ * This is a NOP for replication modes other than ORIGIN.
+ */
+SELVA_IO_EXPORT(void, selva_replication_replicate, int64_t ts, int8_t cmd, const void *buf, size_t buf_size);
+
+/**
+ * Replicate a command to replicas.
+ * Pass the ownership of buf to the replication module. Avoids one malloc.
+ * buf must be allocated with `selva_malloc` or `selva_realloc`.
+ */
+SELVA_IO_EXPORT(void, selva_replication_replicate_pass, int64_t ts, int8_t cmd, void *buf, size_t buf_size);
+
 #define _import_selva_io(apply) \
     apply(selva_io_get_ver) \
+    apply(selva_io_register_serializer) \
+    apply(selva_io_set_dirty) \
+    apply(selva_io_get_dump_state) \
     apply(selva_io_open_last_good) \
     apply(selva_io_last_good_info) \
     apply(selva_io_read_hash) \
@@ -165,7 +237,10 @@ SELVA_IO_EXPORT(struct selva_string *, selva_io_load_string, struct selva_io *io
     apply(selva_io_load_signed) \
     apply(selva_io_load_double) \
     apply(selva_io_load_str) \
-    apply(selva_io_load_string)
+    apply(selva_io_load_string) \
+    apply(selva_replication_get_mode) \
+    apply(selva_replication_replicate) \
+    apply(selva_replication_replicate_pass)
 
 #define _import_selva_io1(f) \
     evl_import(f, "mod_io.so");