+++ selvad/modules/io/io.c
@@ -10,24 +10,37 @@
 #include <stdint.h>
 #include <stdio.h>
 #include <string.h>
+#include <time.h>
 #include <unistd.h>
 #include "endian.h"
 #include "jemalloc.h"
 #include "sha3iuf/sha3.h"
 #include "util/selva_string.h"
+#include "config.h"
 #include "module.h"
+#include "event_loop.h"
+#include "evl_signal.h"
 #include "selva_error.h"
 #include "selva_log.h"
 #include "selva_proto.h"
 #include "selva_io.h"
+#include "selva_server.h"
 #include "myreadlink.h"
 #include "sdb.h"
+#include "dump.h"
+#include "replication/replication.h"
 
 #define READ_WHENCE_READ_HEADER "read header"
 #define READ_WHENCE_HEADER_TYPE "type"
 #define READ_WHENCE_VALUE       "read value"
 
 static const char last_good_name[] = "dump.sdb";
+
+static const struct config cfg_map[] = {
+    { "SELVA_REPLICATION_MODE", CONFIG_INT, &replication_mode },
+    { "SAVE_AT_EXIT",           CONFIG_INT, &save_at_exit },
+    { "AUTO_SAVE_INTERVAL",     CONFIG_INT, &auto_save_interval },
+};
 
 /**
  * Test that the IO mode is set as expected.
@@ -492,10 +505,22 @@ struct selva_string *selva_io_load_string(struct selva_io *io)
 }
 
 IMPORT() {
+	evl_import_signal();
+    evl_import_event_loop();
+    evl_import_main(config_resolve);
+    evl_import_main(evl_set_timeout);
     evl_import_main(selva_log);
+    import_selva_server();
 }
 
 __constructor static void init(void)
 {
     evl_module_init("io");
+
+	if (config_resolve("io", cfg_map, num_elem(cfg_map))) {
+        exit(EXIT_FAILURE);
+    }
+
+    dump_init();
+    replication_init();
 }