+++ selvad/modules/io/replication/origin/replication.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 SAULX
+ * Copyright (c) 2023-2024 SAULX
  * SPDX-License-Identifier: MIT
  */
 #define _GNU_SOURCE
@@ -9,19 +9,22 @@
 #include <stdatomic.h>
 #include <stddef.h>
 #include <stdint.h>
+#include <stdio.h>
 #include <string.h>
 #include <unistd.h>
+#include "sha3iuf/sha3.h"
 #include "jemalloc.h"
 #include "util/selva_string.h"
 #include "util/timestamp.h"
 #include "event_loop.h"
 #include "selva_error.h"
 #include "selva_log.h"
-#include "selva_replication.h"
+#include "selva_io.h"
 #include "selva_server.h"
 #include "../eid.h"
 #include "ring_buffer.h"
 #include "../replication.h"
+#include "../../../../tunables.h"
 #include "replica.h"
 
 static_assert(sizeof(ring_buffer_eid_t) >= sizeof(uint64_t));